+++ plan-comp/src/main/java/com/tissue/plan/web/spring/controllers/MessageReplyController.java
@@ -51,10 +51,11 @@ public class MessageReplyController {
     @RequestMapping(value="/messages/{messageId}/messageReplies/_create", method=POST)
     public String addMessageReply(@PathVariable("messageId") Message message, @Valid MessageReplyForm form, BindingResult resutl, Map model, @ModelAttribute("viewerAccount") Account viewerAccount) {
 
-        /**
-        Topic topic = message.getArticle().getPlan().getTopic();
-        topicService.checkMember(topic, viewerAccount, model);
-        */
+        if(message == null) {
+            return "accessDenied";
+        }
+
+        //todo: process on binding error
 
         form.setMessage(message);
         form.setAccount(viewerAccount);
@@ -64,29 +65,15 @@ public class MessageReplyController {
     }
 
     /**
-     * Update a PostMessageComment.
-     * The post type can be 'concept', 'note' or 'tutorial'.
-    @RequestMapping(value="/messageReplies/{replyId}/_update", method=POST)
-    public String updateMessageReply(@PathVariable("replyId") MessageReply messageReply, @Valid ContentForm form, BindingResult result, Map model, @ModelAttribute("viewerAccount") Account viewerAccount) {
-
-        form.setId(messageReply.getId());
-        messageReplyService.updateContent(form);
-
-        return "redirect:/articles/" + messageReply.getMessage().getArticle().getId().replace("#", "");
-    }
-     */
-
-    /**
      * Delete a PostMessageComment.
      * The post type can be 'concept', 'note' or 'tutorial'.
      */
     @RequestMapping(value="/messageReplies/{replyId}/_delete", method=POST)
     public String deleteMessageReply(@PathVariable("replyId") MessageReply messageReply, Map model, @ModelAttribute("viewerAccount") Account viewerAccount) {
 
-        /**
-        Topic topic = messageReply.getMessage().getArticle().getPlan().getTopic();
-        topicService.checkMember(topic, viewerAccount, model);
-        */
+        if((messageReply == null) || messageReply.getAccount().getId().equals(viewerAccount.getId())) {
+            return "accessDenied";
+        }
 
         messageReplyService.deleteContent(messageReply.getId());
         return "redirect:/articles/" + messageReply.getMessage().getArticle().getId().replace("#", "");
